requirements.txt
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
chromadb
python-dotenv
langgraph
